Change to updated RangeText class
